Take constants into account
Snooping in stashes to see whether there is a typeglob with a CODE
entry will fail for constants created by ‘use constant’, which are
stored shorthand as scalar references.

In addition, perl 5.28 will introduce an optimisation whereby most
subs are stored as code refs in stashes, without needing a typeglob.
This commit gets all tests passing with bleadperl.